refactor: extract worker-side composition root into worker_bootstrap.py
_indexing/worker_pool.py mixed two concerns: the parent-side
WorkerPoolManager (executor lifecycle) and the worker-side entry points
(_process_file_worker, _init_worker, the process-local _worker_analyzer)
that only ever run inside spawned child processes.

The worker function also instantiated the top-level CppAnalyzer facade
from inside the _indexing layer, inverting the intended dependency
direction (cpp_analyzer -> composition_root -> _indexing).

Move the worker-side code into a new top-level worker_bootstrap.py, the
designated per-process composition root: a live CppAnalyzer cannot be
pickled across the spawn boundary, so each worker must rebuild its own
object graph, and Clean Architecture permits naming concretions in a
composition root. worker_pool.py now contains only WorkerPoolManager.

The CppAnalyzer import stays deferred inside _process_file_worker():
worker_bootstrap sits on the composition_root -> indexing_task_submitter
import chain, so a module-level facade import would be circular in the
parent process. The _worker_analyzer global is now annotated
Optional[CppAnalyzer] via TYPE_CHECKING, matching the codebase idiom.

No behavior change.
